backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-pptx==1.0.2
Pillow==11.1.0
sqlalchemy==2.0.36
python-multipart==0.0.19
pdf2image==1.17.0
